backend/requirements.txt
asgiref
Django
# For Cross Origin request issues
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv